+++ src/container/bodyPanel/routePanels/home/titlePanel/TitlePanel.css
@@ -1,7 +1,7 @@
 .titlePanel{
     width: 100%;
     height: 100%;
-    background-image: url('../../static/images/background.jpg');
+    background-image: url('../../../../../static/images/background.jpg');
     background-position: center; /* Center the image */
     background-repeat: no-repeat; /* Do not repeat the image */
     background-size: cover;
