+++ src/container/bodyPanel/routePanels/home/titlePanel/TitlePanel.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import './TitlePanel.css';
-import MyCard from '../myCard/MyCard';
-import Title from '../title/Title';
+import MyCard from './myCard/MyCard';
+import Title from './title/Title';
 
 export default function TitlePanel() {
   return (
