+++ src/container/bodyPanel/routePanels/home/titlePanel/myCard/MyCard.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import titleImage from '../../static/images/titleImage.jpg';
+import titleImage from '../../../../../..//static/images/titleImage.jpg';
 import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import CardActionArea from '@material-ui/core/CardActionArea';
