+++ src/container/bodyPanel/sideBar/SideBar.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './SideBar.css';
 import Link from '@material-ui/core/Link'
-import Navigation from '../navigation/Navigation';
-import Logo from '../logo/Logo';
+import Navigation from './navigation/Navigation';
+import Logo from './logo/Logo';
 
 export default function SideBar() {
