+++ src/container/bodyPanel/sideBar/TopAppBar.js
@@ -4,8 +4,8 @@ import Drawer from '@material-ui/core/Drawer';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar'
 import Button from '@material-ui/core/Button'
-import Navigation from '../navigation/Navigation'
-import Logo from '../logo/Logo'
+import Navigation from './navigation/Navigation'
+import Logo from './logo/Logo'
 import './SideBar.css'
 
 const useStyles = makeStyles({
