+++ src/container/bodyPanel/sideBar/logo/Logo.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import './Logo.css';
-import logo from '../../static/images/logo.png'
+import logo from '../../../../static/images/logo.png';
 import Image from 'react-bootstrap/Image'
 
 export default function Logo() {
