+++ src/static/routes/routes.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Home from '../../container/home/Home';
-import About from '../../container/about/About'
+import Home from '../../container/bodyPanel/routePanels/home/Home';
+import About from '../../container/bodyPanel/routePanels/home/about/About'
 
 export default [
     {
